Group Trigger stuff streams (but still does nothing) (#1407)
* Group Trigger stuff streams (but still does nothing)

Group trigger streaming in place. Its still all nulls
all the time, isn't queried, doesn't evaluate, etc...
mostly because I ddin't code much today but lets commit
this for the weekend

Streaming also meant a split from operator and storage,
just like LFO etc... and so this puts the Storage idiom in place

* fx

+++ src/utils.h
@@ -314,7 +314,7 @@ template <typename E, std::string (*F)(const E &)>
 inline std::unordered_map<std::string, E> makeEnumInverse(const E &from, const E &to)
 {
     std::unordered_map<std::string, E> res;
-    for (auto i = (int32_t)from; i <= to; ++i)
+    for (auto i = (int32_t)from; i <= (int32_t)to; ++i)
     {
         res[F((E)i)] = (E)i;
     }
